+++ src/main/java/ru/legionofone/klassikaplusserver/service/StockItemUpdateService.java
@@ -6,7 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import ru.legionofone.klassikaplusserver.model.persistance.entities.DbItem;
-import ru.legionofone.klassikaplusserver.model.repository.GenericHibernateDao;
+import ru.legionofone.klassikaplusserver.model.persistance.GenericHibernateDao;
+import ru.legionofone.klassikaplusserver.web.controller.CatalogItemReceiver;
 
 @Service
 public class StockItemUpdateService {
@@ -14,15 +15,18 @@ public class StockItemUpdateService {
     private static final Logger logger = LoggerFactory.getLogger(CatalogService.class);
 
     final GenericHibernateDao<DbItem> itemRepository;
+    final CatalogItemReceiver catalogItemReceiver;
 
     @Autowired
-    public StockItemUpdateService(GenericHibernateDao<DbItem> itemRepository) {
+    public StockItemUpdateService(GenericHibernateDao<DbItem> itemRepository, CatalogItemReceiver catalogItemReceiver) {
         this.itemRepository = itemRepository;
+        this.catalogItemReceiver = catalogItemReceiver;
     }
 
     @Scheduled(fixedRate = 5000)
-    public void updateCatalogItem(){
+    public void updateCatalogItem() {
 
+        catalogItemReceiver.provide();
         //TODO Запрос на сторону сайта
     }
 }
